chore: Adjust nullability, adjust properties visibility
