+++ src/Uno.UI.RemoteControl/HotReload/ClientHotReloadProcessor.Xaml.cs
@@ -87,7 +87,7 @@ namespace Uno.UI.RemoteControl.HotReload
 
                 Application.RegisterComponent(uri, fileContent);
 
-                foreach (var instance in EnumerateInstances(Window.Current.Content, i => uri.OriginalString == i.DebugParseContext.LocalFileUri))
+                foreach (var instance in EnumerateInstances(Window.Current.Content, i => uri.OriginalString == i.DebugParseContext?.LocalFileUri))
                 {
                     switch (instance)
                     {
@@ -213,7 +213,7 @@ namespace Uno.UI.RemoteControl.HotReload
 			
             newView.SetBaseUri(
 				oldView.BaseUri.OriginalString,
-				oldView.DebugParseContext?.LocalFileUri,
+				oldView.DebugParseContext?.LocalFileUri ?? "",
 				oldView.DebugParseContext?.LineNumber ?? -1,
 				oldView.DebugParseContext?.LinePosition ?? -1);
